samples/ingestion-pipeline/requirements.txt
docling
azure-storage-blob 
azure-identity